+++ cardinal-components-item/src/main/java/dev/onyxstudios/cca/internal/item/StaticItemComponentPlugin.java
@@ -22,16 +22,16 @@
  */
 package dev.onyxstudios.cca.internal.item;
 
+import dev.onyxstudios.cca.api.v3.component.ComponentContainer;
 import dev.onyxstudios.cca.api.v3.component.ComponentKey;
-import dev.onyxstudios.cca.api.v3.component.item.ItemComponentFactory;
-import dev.onyxstudios.cca.api.v3.component.item.ItemComponentFactoryRegistry;
-import dev.onyxstudios.cca.api.v3.component.item.ItemComponentFactoryV2;
-import dev.onyxstudios.cca.api.v3.component.item.ItemComponentInitializer;
+import dev.onyxstudios.cca.api.v3.item.ItemComponentFactory;
+import dev.onyxstudios.cca.api.v3.item.ItemComponentFactoryRegistry;
+import dev.onyxstudios.cca.api.v3.item.ItemComponentFactoryV2;
+import dev.onyxstudios.cca.api.v3.item.ItemComponentInitializer;
 import dev.onyxstudios.cca.internal.base.LazyDispatcher;
 import dev.onyxstudios.cca.internal.base.asm.CcaAsmHelper;
 import dev.onyxstudios.cca.internal.base.asm.StaticComponentLoadingException;
 import dev.onyxstudios.cca.internal.base.asm.StaticComponentPluginBase;
-import nerdhub.cardinal.components.api.component.ComponentContainer;
 import nerdhub.cardinal.components.api.component.extension.CopyableComponent;
 import nerdhub.cardinal.components.api.event.ItemComponentCallbackV2;
 import net.fabricmc.loader.api.FabricLoader;
@@ -40,15 +40,17 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.util.Identifier;
 import org.jetbrains.annotations.Nullable;
 
+import javax.annotation.Nonnull;
 import java.io.IOException;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Objects;
+import java.util.*;
+import java.util.function.Predicate;
 
 public final class StaticItemComponentPlugin extends LazyDispatcher implements ItemComponentFactoryRegistry {
     public static final StaticItemComponentPlugin INSTANCE = new StaticItemComponentPlugin();
-    public static final String WILDARD_IMPL_SUFFIX = "ItemStackImpl_All";
+    public static final String WILDCARD_IMPL_SUFFIX = "ItemStackImpl_All";
+    private static final boolean DEV = Boolean.getBoolean("fabric.development");
+    // TODO enable by default when Component#isComponentEqual is gone
+    private static final boolean ENABLE_CHECKS = Boolean.getBoolean("cca.debug.verifyequals");
 
     private StaticItemComponentPlugin() {
         super("creating an ItemStack");
@@ -58,16 +60,30 @@ public final class StaticItemComponentPlugin extends LazyDispatcher implements I
         return "ItemStackImpl_" + CcaAsmHelper.getJavaIdentifierName(itemId);
     }
 
+    private final List<PredicatedComponentFactory<?>> dynamicFactories = new ArrayList<>();
     private final Map<@Nullable Identifier, Map</*ComponentType*/Identifier, ItemComponentFactoryV2<?>>> componentFactories = new HashMap<>();
-    private final Map<Identifier, Class<? extends ItemComponentContainerFactory>> factoryClasses = new HashMap<>();
     private Class<? extends ItemComponentContainerFactory> wildcardFactoryClass;
 
-    public Class<? extends ItemComponentContainerFactory> getFactoryClass(Identifier itemId) {
+    public Class<? extends ItemComponentContainerFactory> getFactoryClass(Item item, Identifier itemId) {
         this.ensureInitialized();
-        Class<? extends ItemComponentContainerFactory> specificFactory = this.factoryClasses.get(itemId);
-        if (specificFactory != null) {
-            return specificFactory;
+        Objects.requireNonNull(item);
+
+        for (PredicatedComponentFactory<?> dynamicFactory : this.dynamicFactories) {
+            dynamicFactory.tryRegister(item, itemId);
         }
+
+        if (this.componentFactories.containsKey(itemId)) {
+            try {
+                Map<Identifier, ItemComponentFactoryV2<?>> compiled = new HashMap<>(this.componentFactories.get(itemId));
+                this.getWildcard().forEach(compiled::putIfAbsent);
+                String implSuffix = getSuffix(itemId);
+                Class<? extends ComponentContainer<CopyableComponent<?>>> containerCls = StaticComponentPluginBase.spinComponentContainer(ItemComponentFactoryV2.class, CopyableComponent.class, compiled, implSuffix);
+                return StaticComponentPluginBase.spinContainerFactory(implSuffix, ItemComponentContainerFactory.class, containerCls, ItemComponentCallbackV2.class, 2, Item.class, ItemStack.class);
+            } catch (IOException e) {
+                throw new StaticComponentLoadingException("Failed to generate a dedicated component container for " + itemId, e);
+            }
+        }
+
         assert this.wildcardFactoryClass != null;
         return this.wildcardFactoryClass;
     }
@@ -78,25 +94,17 @@ public final class StaticItemComponentPlugin extends LazyDispatcher implements I
             FabricLoader.getInstance().getEntrypointContainers("cardinal-components-item", ItemComponentInitializer.class),
             initializer -> initializer.registerItemComponentFactories(this)
         );
-        Map<Identifier, ItemComponentFactoryV2<?>> wildcardMap = this.componentFactories.getOrDefault(null, Collections.emptyMap());
+
         try {
-            Class<? extends ComponentContainer<?>> containerCls = StaticComponentPluginBase.spinComponentContainer(ItemComponentFactoryV2.class, wildcardMap, WILDARD_IMPL_SUFFIX);
-            this.wildcardFactoryClass = StaticComponentPluginBase.spinContainerFactory(WILDARD_IMPL_SUFFIX, ItemComponentContainerFactory.class, containerCls, ItemComponentCallbackV2.class, 2, Item.class, ItemStack.class);
+            Class<? extends ComponentContainer<CopyableComponent<?>>> containerCls = StaticComponentPluginBase.spinComponentContainer(ItemComponentFactoryV2.class, CopyableComponent.class, this.getWildcard(), WILDCARD_IMPL_SUFFIX);
+            this.wildcardFactoryClass = StaticComponentPluginBase.spinContainerFactory(WILDCARD_IMPL_SUFFIX, ItemComponentContainerFactory.class, containerCls, ItemComponentCallbackV2.class, 2, Item.class, ItemStack.class);
         } catch (IOException e) {
             throw new StaticComponentLoadingException("Failed to generate the fallback component container for item stacks", e);
         }
-        for (Map.Entry<Identifier, Map<Identifier, ItemComponentFactoryV2<?>>> entry : this.componentFactories.entrySet()) {
-            if (entry.getKey() == null) continue;
-            try {
-                Map<Identifier, ItemComponentFactoryV2<?>> compiled = new HashMap<>(entry.getValue());
-                wildcardMap.forEach(compiled::putIfAbsent);
-                String implSuffix = getSuffix(entry.getKey());
-                Class<? extends ComponentContainer<?>> containerCls = StaticComponentPluginBase.spinComponentContainer(ItemComponentFactoryV2.class, compiled, implSuffix);
-                this.factoryClasses.put(entry.getKey(), StaticComponentPluginBase.spinContainerFactory(implSuffix, ItemComponentContainerFactory.class, containerCls, ItemComponentCallbackV2.class, 2, Item.class, ItemStack.class));
-            } catch (IOException e) {
-                throw new StaticComponentLoadingException("Failed to generate a dedicated component container for " + entry.getKey(), e);
-            }
-        }
+    }
+
+    private Map<Identifier, ItemComponentFactoryV2<?>> getWildcard() {
+        return this.componentFactories.getOrDefault(null, Collections.emptyMap());
     }
 
     @Override
@@ -121,13 +129,74 @@ public final class StaticItemComponentPlugin extends LazyDispatcher implements I
         this.register(itemId, type, factory);
     }
 
+    @Override
+    public <C extends CopyableComponent<?>> void registerFor(Predicate<Item> test, ComponentKey<? super C> type, ItemComponentFactoryV2<C> factory) {
+        this.dynamicFactories.add(new PredicatedComponentFactory<>(test, type, factory));
+    }
+
     private <C extends CopyableComponent<?>> void register(@Nullable Identifier itemId, ComponentKey<? super C> type, ItemComponentFactoryV2<C> factory) {
         this.checkLoading(ItemComponentFactoryRegistry.class, "register");
+        this.register0(itemId, type, factory);
+    }
+
+    private <C extends CopyableComponent<?>> void register0(@Nullable Identifier itemId, ComponentKey<? super C> type, ItemComponentFactoryV2<C> factory) {
         Map<Identifier, ItemComponentFactoryV2<?>> specializedMap = this.componentFactories.computeIfAbsent(itemId, t -> new HashMap<>());
         ItemComponentFactoryV2<?> previousFactory = specializedMap.get(type.getId());
         if (previousFactory != null) {
             throw new StaticComponentLoadingException("Duplicate factory declarations for " + type.getId() + " on " + (itemId == null ? "every item" : "item '" + itemId + "'") + ": " + factory + " and " + previousFactory);
         }
-        specializedMap.put(type.getId(), (item, stack) -> Objects.requireNonNull(((ItemComponentFactoryV2<?>) factory).createForStack(item, stack), "Component factory "+ factory + " for " + type.getId() + " returned null on " + stack));
+        ItemComponentFactoryV2<CopyableComponent<?>> nonnullFactory = (item, stack) -> Objects.requireNonNull(
+            ((ItemComponentFactoryV2<?>) factory).createForStack(item, stack),
+            "Component factory " + factory + " for " + type.getId() + " returned null on " + stack
+        );
+        ItemComponentFactoryV2<CopyableComponent<?>> finalFactory;
+
+        if (DEV && ENABLE_CHECKS) {
+            finalFactory = new ItemComponentFactoryV2<CopyableComponent<?>>() {
+                private boolean checked;
+
+                @Nonnull
+                @Override
+                public CopyableComponent<?> createForStack(Item item, ItemStack stack) {
+                    CopyableComponent<?> component = nonnullFactory.createForStack(item, stack);
+
+                    if (!this.checked) {
+                        try {
+                            if (component.getClass().getMethod("equals", Object.class).getDeclaringClass() == Object.class) {
+                                throw new IllegalStateException("Component implementation " + component.getClass().getTypeName() + " attached to " + stack + " does not override Object#equals");
+                            }
+                        } catch (NoSuchMethodException e) {
+                            throw new AssertionError("Object#equals not found ?!");
+                        }
+
+                        this.checked = true;
+                    }
+
+                    return component;
+                }
+            };
+        } else {
+            finalFactory = nonnullFactory;
+        }
+
+        specializedMap.put(type.getId(), finalFactory);
+    }
+
+    private final class PredicatedComponentFactory<C extends CopyableComponent<?>> {
+        private final Predicate<Item> predicate;
+        private final ComponentKey<? super C> type;
+        private final ItemComponentFactoryV2<C> factory;
+
+        public PredicatedComponentFactory(Predicate<Item> predicate, ComponentKey<? super C> type, ItemComponentFactoryV2<C> factory) {
+            this.type = type;
+            this.factory = factory;
+            this.predicate = predicate;
+        }
+
+        public void tryRegister(Item item, Identifier id) {
+            if (this.predicate.test(item)) {
+                StaticItemComponentPlugin.this.register0(id, this.type, this.factory);
+            }
+        }
     }
 }