+++ changelog.md
@@ -1,4 +1,34 @@
 ------------------------------------------------------
+Version 2.5.0
+------------------------------------------------------
+**This release breaks compatibility between modules of different versions.**
+- Updated to 1.16.2
+
+Additions
+- Added a new [API for block components](https://github.com/OnyxStudios/Cardinal-Components-API/wiki/Cardinal-Components-Block) (see PR #42)
+- Added a new API for scoreboard components (see PR #36)
+- Added reworked `ComponentProvider` and `ComponentContainer` interfaces to the V3 API
+    - Refactored a lot of internal code to use the new types
+- Added `ComponentV3` and `ComponentRegistryV3` interfaces for early V3 adopters
+- Added a factory builder to the `ComponentContainer` interface for third-party `ComponentProvider` implementations
+- Added alternatives to `registerFor` methods taking a predicate, for fine-tuned registration
+
+Changes
+- The NBT serialized form of `ComponentContainer` is now more space-efficient.
+    - Instances serialized before this update should get converted as they load.
+- Removed experimental deprecated method `EntityComponentFactoryRegistry#register` (use `registerFor`)
+
+Fixes
+- Fixed item components not being copied by the smithing table
+- Removed the `Internal` annotation from new API packages
+
+------------------------------------------------------
+Version 2.4.2
+------------------------------------------------------
+- Fixed regression in `cardinal-components-item` causing crash with Optifine installed
+- Marked the `cardinal-components-level` module as incompatible with 1.16.2 and above
+
+------------------------------------------------------
 Version 2.4.1
 ------------------------------------------------------
 - Fixed crash when more than 16 components were registered